memcg: add mem_cgroup_mark_inode_dirty()
Create the mem_cgroup_mark_inode_dirty() routine, which is called when
an inode is marked dirty.  In kernels without memcg, this is an inline
no-op.

Add i_memcg field to struct address_space.  When an inode is marked
dirty with mem_cgroup_mark_inode_dirty(), the css_id of current memcg is
recorded in i_memcg.  Per-memcg writeback (introduced in a latter
change) uses this field to isolate inodes associated with a particular
memcg. The type of i_memcg is an 'unsigned short' because it stores
the css_id of the memcg.

A mem_cgroup_mark_inode_dirty() tracepoint is also included to allow for
easier understanding of memcg writeback operation.

Ref: https://lkml.org/lkml/2011/8/17/200

Signed-off-by: Sha Zhengju <handai.szj@taobao.com>